+++ R/report_sample.R
@@ -9,6 +9,29 @@
 #'   calculated for numeric variables. May be `"mean"` (for mean and
 #'   standard deviation) or `"median"` (for median and median absolute
 #'   deviation) as summary.
+#' @param ci Level of confidence interval for relative frequencies (proportions).
+#'   If not `NULL`, confidence intervals are shown for proportions of factor
+#'   levels. If `p` is the proportion of a factor level, the related confidence
+#'   interval is calculated as follows:
+#'
+#'   \deqn{p +/- z \sqrt{\frac{p (1 - p)}{n}}}
+#'
+#'   where `z` is the critical z-score based on the interval level and `n` is
+#'   the length of the vector.
+#' @param ci_adjust Scalar, if not `NULL`, applies *Wilson's adjustment* to
+#'   confidence intervals of proportions that are close to 0 or 1. `ci_adjust`
+#'   must be a value between 0 and 1, indicating how close a proportions must
+#'   be to 0 or 1 in order to apply the adjustment to its confidence intervals.
+#'   E.g., if `ci_adjust = 0.07`, confidence intervals of all proportions smaller
+#'   than 0.07 or greater than 0.93 will be adjusted. Their calculation is as
+#'   follows:
+#'
+#'   \deqn{p +/- z \sqrt{\frac{p (1 - p)}{n + 4}}}
+#'
+#'   where `p` is
+#'
+#'   \deqn{p = \frac{x + 2}{n + 4}}
+#'
 #' @param select Character vector, with column names that should be included in
 #'   the descriptive table.
 #' @param exclude Character vector, with column names that should be excluded
@@ -25,6 +48,10 @@
 #' @return A data frame of class `report_sample` with variable names and
 #'   their related summary statistics.
 #'
+#' @references
+#' Wilson, E. B. (1927). Probable inference, the law of succession, and statistical
+#' inference. Journal of the American Statistical Association. 22 (158): 209–212
+#'
 #' @examples
 #' library(report)
 #'
@@ -36,6 +63,8 @@
 report_sample <- function(data,
                           group_by = NULL,
                           centrality = "mean",
+                          ci = NULL,
+                          ci_adjust = NULL,
                           select = NULL,
                           exclude = NULL,
                           weights = NULL,
@@ -71,7 +100,7 @@ report_sample <- function(data,
   out <- if (isTRUE(grouping)) {
     result <- lapply(split(data[variables], factor(data[[group_by]])), function(x) {
       x[[group_by]] <- NULL
-      .generate_descriptive_table(x, centrality, weights, digits, n)
+      .generate_descriptive_table(x, centrality, weights, digits, n, ci, ci_adjust)
     })
     # remember values of first columns
     variable <- result[[1]]["Variable"]
@@ -97,7 +126,9 @@ report_sample <- function(data,
         centrality,
         weights,
         digits,
-        n
+        n,
+        ci,
+        ci_adjust
       )[["Summary"]]
     )
     # Remove Total column if need be
@@ -118,7 +149,7 @@ report_sample <- function(data,
     )
     final
   } else {
-    .generate_descriptive_table(data[variables], centrality, weights, digits, n)
+    .generate_descriptive_table(data[variables], centrality, weights, digits, n, ci, ci_adjust)
   }
 
   attr(out, "weighted") <- !is.null(weights)
@@ -129,7 +160,13 @@ report_sample <- function(data,
 
 # helper ------------------------
 
-.generate_descriptive_table <- function(x, centrality, weights, digits, n = FALSE) {
+.generate_descriptive_table <- function(x,
+                                        centrality,
+                                        weights,
+                                        digits,
+                                        n = FALSE,
+                                        ci = NULL,
+                                        ci_adjust = NULL) {
   if (!is.null(weights)) {
     w <- x[[weights]]
     columns <- setdiff(colnames(x), weights)
@@ -145,7 +182,9 @@ report_sample <- function(data,
       centrality = centrality,
       weights = w,
       digits = digits,
-      n = n
+      n = n,
+      ci = ci,
+      ci_adjust = ci_adjust
     )
   }))
 }
@@ -198,6 +237,8 @@ report_sample <- function(data,
                                       column,
                                       weights = NULL,
                                       digits = 1,
+                                      ci = NULL,
+                                      ci_adjust = NULL,
                                       ...) {
   if (!is.null(weights)) {
     x[is.na(weights)] <- NA
@@ -213,9 +254,22 @@ report_sample <- function(data,
   if (length(proportions) == 2) {
     proportions <- proportions[2]
   }
-  .summary <- vapply(proportions, function(i) sprintf("%.1f", 100 * i), "character")
+
+  # CI for proportions?
+  if (!is.null(ci)) {
+    se <- .se_proportion(x, proportions, weights, ci, ci_adjust)
+    .summary <- sprintf(
+      "%.1f (%.1f, %.1f)",
+      100 * proportions,
+      100 * (proportions - se),
+      100 * (proportions + se)
+    )
+  } else {
+    .summary <- sprintf("%.1f", 100 * proportions)
+  }
+
   data.frame(
-    Variable = sprintf("%s [%s], %%", column, names(.summary)),
+    Variable = sprintf("%s [%s], %%", column, names(proportions)),
     Summary = as.vector(.summary),
     stringsAsFactors = FALSE
   )
@@ -226,8 +280,47 @@ report_sample <- function(data,
 
 
 
-# print-method --------------------------------------------
+# Standard error for confidence interval of proportions ----
 
+.se_proportion <- function(x, proportions, weights, ci, ci_adjust) {
+  p_hat <- proportions
+  n <- length(stats::na.omit(x))
+
+  # Wilson's adjustment for p close to 0 or 1
+  if (!is.null(ci_adjust) && !is.na(ci_adjust) && is.numeric(ci_adjust)) {
+    fix <- p_hat < ci_adjust | p_hat > (1 - ci_adjust)
+    if (any(fix)) {
+      if (is.null(weights)) {
+        p_adjusted <- (as.vector(table(x)) + 2) / (n + 4)
+      } else {
+        p_adjusted <- (as.vector(round(stats::xtabs(weights ~ x))) + 2) / (round(sum(weights)) + 4)
+      }
+      # for binary factors, just need one level
+      if (length(p_adjusted) == 2) {
+        p_adjusted <- p_adjusted[2]
+      }
+      p_hat[fix] <- p_adjusted[fix]
+      n <- rep(n, length(p_hat))
+      n[fix] <- n[fix] + 4
+    }
+  }
+
+  ## FIXME: Once know how to do, calculate accurate SE for weighted data
+
+  # There is a formula how to deal with SE for weighted data, see
+  # https://en.wikipedia.org/wiki/Binomial_proportion_confidence_interval
+  # but it seems "p_hat" is unknown. For now, we give a warning instead of
+  # estimating p_hat for weighted data
+  if (!is.null(weights)) {
+    insight::format_warning("Confidence intervals are not accurate for weighted data.")
+  }
+
+  stats::qnorm((1 + ci) / 2) * suppressWarnings(sqrt(p_hat * (1 - p_hat) / n))
+}
+
+
+
+# print-method --------------------------------------------
 
 #' @export
 print.report_sample <- function(x, ...) {