+++ tests/testthat/test-report_sample.R
@@ -69,6 +69,32 @@ test_that("report_sample n = TRUE", {
   )
 })
 
+test_that("report_sample CI", {
+  expect_snapshot(
+    variant = "windows",
+    report_sample(iris, select = c("Sepal.Length", "Species"), ci = 0.95)
+  )
+
+  set.seed(123)
+  d <- data.frame(
+    x = as.factor(rbinom(1000, 1, prob = 0.03)),
+    w = abs(rnorm(1000, 1, 0.1))
+  )
+  expect_snapshot(
+    variant = "windows",
+    report_sample(d, ci = 0.95, select = "x")
+  )
+  expect_snapshot(
+    variant = "windows",
+    report_sample(d, ci = 0.95, ci_adjust = 0.04, select = "x")
+  )
+  expect_snapshot(
+    variant = "windows",
+    report_sample(d, ci = 0.95, ci_adjust = 0.02, select = "x")
+  )
+  expect_warning(report_sample(d, ci = 0.95, weights = "w"), regex = "accurate")
+})
+
 test_that("report_sample group_by", {
   expect_snapshot(
     variant = "windows",